// src/tmdb/models/tmdb.models.ts
export type TmdbMovie = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
};

export type Movie = {
    title: string;
    id: string;
    language: string;
    release_date: string;
    overview: string;
}